fixed route name issue on user controller, renamed authRouter to userRouter
