+++ src/index.ts
@@ -1,5 +1,5 @@
 import app from './app';
 
-app.listen(80, () => {
+app.listen(9000, () => {
 	console.log('Tenant API listening for requests...');
 });
